pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "eda-clustering-time-series-regression-model-airbnb-data"
version = "0.1.0"
description = "EDA, clustering, and time series regression model pipeline for Airbnb data"
readme = "README.md"
requires-python = ">=3.11"

# Dependencies
dependencies = [
    "pyodbc",
    "sqlalchemy",
    "pandas",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "python-dotenv"
]

# Optional metadata
authors = [{ name = "Yassine" }]
license = { text = "MIT" }

[tool.setuptools.packages.find]
where = ["."]
